Rename RayDiNamedAnnotation to AnnotationBinding
