+++ rector.php
@@ -2,19 +2,12 @@
 
 declare(strict_types=1);
 
-use Rector\Ray\RayDiNamedAnnotation\Rector\ClassMethod\RayDiNamedAnnotationRector;
-use Rector\Core\Configuration\Option;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Rector\Config\RectorConfig;
+use Rector\Ray\RayDiNamedAnnotation\Rector\ClassMethod\AnnotationBindingRector;
 
 require __DIR__ . '/vendor/autoload.php';
 
-return static function (ContainerConfigurator $containerConfigurator): void {
-    // get parameters
-    $parameters = $containerConfigurator->parameters();
-    $parameters->set(Option::PATHS, [
-        __DIR__ . '/src'
-    ]);
-
-    $services = $containerConfigurator->services();
-    $services->set(RayDiNamedAnnotationRector::class);
+return static function (RectorConfig $config): void {
+    $services = $config->services();
+    $services->set(AnnotationBindingRector::class);
 };
