+++ rules-tests/AnnoattionBinding/Rector/ClassMethod/AnnotationBindingRector/Fixture/Bar.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\AnnoattionBinding\Rector\ClassMethod\AnnotationBindingRector\Fixture;
+
+use Ray\Di\Di\Qualifier;
+
+/**
+ * @Annotation
+ * @Qualifier
+ */
+class Bar
+{
+}
