+++ rules-tests/AnnoattionBinding/Rector/ClassMethod/AnnotationBindingRector/Fixture/Foo.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\AnnoattionBinding\Rector\ClassMethod\AnnotationBindingRector\Fixture;
+
+use Ray\Di\Di\Qualifier;
+
+/**
+ * @Annotation
+ * @Qualifier
+ */
+class Foo
+{
+    public string $value = '';
+}
