+++ rules-tests/AnnoattionBinding/Rector/ClassMethod/AnnotationBindingRector/Fixture/QualifierNoValue.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Tests\AnnoattionBinding\Rector\ClassMethod\AnnotationBindingRector\Fixture;
+
+class QualifierNoValue
+{
+    /** @Bar */
+    public function __construct(int $a, int $b)
+    {
+    }
+}
