+++ rules-tests/AnnoattionBinding/Rector/ClassMethod/AnnotationBindingRector/Fixture/quolifier_no_value.php.inc
@@ -1,0 +1,28 @@
+<?php
+
+namespace Rector\Tests\AnnoattionBinding\Rector\ClassMethod\AnnotationBindingRector\Fixture;
+
+class QualifierNoValue
+{
+    /**
+     * @Bar
+     */
+    public function __construct(int $a, int $b)
+    {
+    }
+}
+?>
+-----
+<?php
+
+namespace Rector\Tests\AnnoattionBinding\Rector\ClassMethod\AnnotationBindingRector\Fixture;
+
+class QualifierNoValue
+{
+    public function __construct(#[\Rector\Tests\AnnoattionBinding\Rector\ClassMethod\AnnotationBindingRector\Fixture\Bar]
+    int $a, #[\Rector\Tests\AnnoattionBinding\Rector\ClassMethod\AnnotationBindingRector\Fixture\Bar]
+    int $b)
+    {
+    }
+}
+?>
