+++ rules-tests/AnnoattionBinding/Rector/ClassMethod/AnnotationBindingRector/config/configured_rule.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace RectorPrefix20220323;
 
 use Rector\Config\RectorConfig;
-use Rector\Ray\RayDiNamedAnnotation\Rector\ClassMethod\RayDiNamedAnnotationRector;
+use Rector\Ray\AnnotationBinding\Rector\ClassMethod\AnnotationBindingRector;
 
 return static function (RectorConfig $config): void {
     $services = $config->services();
-    $services->set(RayDiNamedAnnotationRector::class);
+    $services->set(AnnotationBindingRector::class);
 };
