--- rules-tests/RayDiNamedAnnotation/Rector/ClassMethod/RayDiNamedAnnotationRector/Fake/Bar.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Rector\Tests\RayDiNamedAnnotation\Rector\ClassMethod\RayDiNamedAnnotationRector\Fixture;
-
-use Ray\Di\Di\Qualifier;
-
-/**
- * @Annotation
- * @Qualifier
- */
-class Bar
-{
-}
-
