--- rules-tests/RayDiNamedAnnotation/Rector/ClassMethod/RayDiNamedAnnotationRector/Fake/Foo.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Rector\Tests\RayDiNamedAnnotation\Rector\ClassMethod\RayDiNamedAnnotationRector\Fixture;
-
-use Ray\Di\Di\Qualifier;
-
-/**
- * @Annotation
- * @Qualifier
- */
-class Foo
-{
-    public string $value = '';
-}
-
