--- rules-tests/RayDiNamedAnnotation/Rector/ClassMethod/RayDiNamedAnnotationRector/Fake/NoneNamed.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Rector\Tests\RayDiNamedAnnotation\Rector\ClassMethod\RayDiNamedAnnotationRector\Fixture;
-
-class NoneNamed
-{
-    /**
-     * @Foo
-     */
-    public function __construct(int $a, int $b)
-    {
-    }
-}
