--- rules-tests/RayDiNamedAnnotation/Rector/ClassMethod/RayDiNamedAnnotationRector/Fake/QualifierNoValue.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Rector\Tests\RayDiNamedAnnotation\Rector\ClassMethod\RayDiNamedAnnotationRector\Fixture;
-
-class QualifierNoValue
-{
-    /**
-     * @Bar
-     */
-    public function __construct(int $a, int $b)
-    {
-    }
-}
