--- rules-tests/RayDiNamedAnnotation/Rector/ClassMethod/RayDiNamedAnnotationRector/Fake/QualifierParam.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Rector\Tests\RayDiNamedAnnotation\Rector\ClassMethod\RayDiNamedAnnotationRector\Fixture;
-
-class QualifierParam
-{
-    /**
-     * @Foo("a")
-     */
-    public function __construct(int $a, int $b)
-    {
-    }
-}
